+++ src/main/java/nextstep/oauth2/google/GithubAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package nextstep.security.oauth2.github;
+package nextstep.oauth2.google;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -21,13 +21,12 @@ import java.io.IOException;
 import java.util.Map;
 import java.util.Set;
 
-public class GoogleAuthenticationFilter extends OncePerRequestFilter {
-
-    public static final String GOOGLE_AUTHORIZATION_REDIRECT_URI = "/login/oauth2/code/google";
-    public static final String GOOGLE_USER_INFO_REQUEST_URI = "http://localhost:8089/oauth2/v1/userinfo";
-    private static final String GOOGLE_ACCESS_TOKEN_REQUEST_URI = "http://localhost:8089/token";
-    //    public static final String GOOGLE_ACCESS_TOKEN_REQUEST_URI = "https://oauth2.googleapis.com/token";
-//    public static final String GOOGLE_USER_INFO_REQUEST_URI = "https://www.googleapis.com/oauth2/v1/userinfo";
+public class GithubAuthenticationFilter extends OncePerRequestFilter {
+    public static final String GITHUB_AUTHORIZATION_REDIRECT_URI = "/login/oauth2/code/github";
+        public static final String GITHUB_USER_INFO_REQUEST_URI = "http://localhost:8089/user";
+    private static final String GITHUB_ACCESS_TOKEN_REQUEST_URI = "http://localhost:8089/login/oauth/access_token";
+//    public static final String GITHUB_ACCESS_TOKEN_REQUEST_URI = "https://github.com/login/oauth/access_token";
+//    public static final String GITHUB_USER_INFO_REQUEST_URI = "https://api.github.com/user";
     private static final RestClient restClient = RestClient.create();
     private final MemberRepository memberRepository = new InmemoryMemberRepository();
     private final HttpSessionSecurityContextRepository securityContextRepository = new HttpSessionSecurityContextRepository();
@@ -38,18 +37,16 @@ public class GoogleAuthenticationFilter extends OncePerRequestFilter {
 
     @Override
     protected void doFilterInternal(final HttpServletRequest request, final HttpServletResponse response, final FilterChain filterChain) throws ServletException, IOException {
-        if (!request.getRequestURI().equals(GOOGLE_AUTHORIZATION_REDIRECT_URI)) {
+        if (!request.getRequestURI().equals(GITHUB_AUTHORIZATION_REDIRECT_URI)) {
             doFilter(request, response, filterChain);
             return;
         }
 
         final String code = request.getParameter("code");
-        final Map<String, String> tokenResponse = sendPostRequest(GOOGLE_ACCESS_TOKEN_REQUEST_URI, code);
-        System.out.println("tokenResponse = " + tokenResponse);
+        final Map<String, String> tokenResponse = sendPostRequest(GITHUB_ACCESS_TOKEN_REQUEST_URI, code);
 
         final String accessToken = tokenResponse.get("access_token");
-        final Map<String, String> userInfoResponse = sendGetRequestWithToken(GOOGLE_USER_INFO_REQUEST_URI, accessToken);
-        System.out.println("userInfoResponse = " + userInfoResponse);
+        final Map<String, String> userInfoResponse = sendGetRequestWithToken(GITHUB_USER_INFO_REQUEST_URI, accessToken);
 
         final Member member = retrieveMember(userInfoResponse);
 
@@ -70,7 +67,7 @@ public class GoogleAuthenticationFilter extends OncePerRequestFilter {
     private Member retrieveMember(final Map<String, String> userInfoResponse) {
         final String email = userInfoResponse.get("email");
         final String name = userInfoResponse.get("name");
-        final String avatarUrl = userInfoResponse.get("picture");
+        final String avatarUrl = userInfoResponse.get("avatar_url");
 
         final Member member = memberRepository.findByEmail(email)
                 .orElseGet(() -> memberRepository.save(new Member(email, "", name, avatarUrl, Set.of())));
@@ -87,15 +84,13 @@ public class GoogleAuthenticationFilter extends OncePerRequestFilter {
 
     public Map<String, String> sendPostRequest(String url, String code) {
         MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
-        body.add("client_id", "176481963463-r1n0954hb0g0tmrefbiq920k7014ka0n.apps.googleusercontent.com");
+        body.add("client_id", "Ov23liWTGdV9MeJBRLwC");
         body.add("client_secret", "secret");
         body.add("code", code);
-        body.add("redirect_uri", GoogleLoginRedirectFilter.REDIRECT_URI);
-        body.add("grant_type", "authorization_code");
+        body.add("redirect_uri", GithubLoginRedirectFilter.REDIRECT_URI);
 
         return restClient.post()
                 .uri(url)
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED)
                 .accept(MediaType.APPLICATION_JSON)
                 .body(body)
                 .retrieve()
